DataVizpro/requirements.txt
Flask>=2.3.3
Pandas>=2.1.0
Matplotlib>=3.8.0
Seaborn>=0.13.0
Scikit-learn>=1.3.0
fpdf2>=2.7.0
dash>=2.14.0
plotly>=5.18.0
dash-bootstrap-components>=1.5.0
dash-daq>=0.5.0
